refactor static history to use util helpers
